// src/puzzle/randomTemplates.generated.ts
export const RANDOM_PUZZLE_TEMPLATES: Record<string, RandomPuzzleTemplate> = {
  "cube-42": {
    "arrows": [
      {
        "id": "arrow-1",
        "path": [
          [
            8,
            4,
            5
          ],
          [
            5,
            4,
            5
          ],
          [
            5,
            4,
            8
          ],
          [
            2,
            4,
            8
          ],
          [
            2,
            4,
            5
          ]
        ],
        "exitDirection": "-Z",
        "color": 3120282,
        "lengthClass": "long"
      },
// ...
  },
// ...
};